used swiftlint to format files
