+++ Lunar/Kbin/KbinCommentsView.swift
@@ -91,18 +91,9 @@ struct KbinCommentRowView: View {
     }
 }
 
-// struct KbinCommentsView_Previews: PreviewProvider {
-//    static var previews: some View {
-//        /// need to set showing popover to a constant value
-//        KbinCommentsView(
-//            kbinCommentsFetcher: KbinCommentsFetcher(
-//                postURL: MockData.kbinPostURL
-//            ),
-//            kbinThreadBodyFetcher: KbinThreadBodyFetcher(
-//                postURL: MockData.kbinPostURL2
-//            ),
-//            postURL: MockData.kbinPostURL,
-//            post: MockData.kbinPost
-//        )
-//    }
-// }
+ struct KbinCommentsView_Previews: PreviewProvider {
+    static var previews: some View {
+        /// need to set showing popover to a constant value
+        KbinCommentsView(postURL: MockData.kbinPostURL)
+    }
+ }
