+++ Lunar/Login Views/UsernameFieldView.swift
@@ -33,7 +33,8 @@ struct UsernameFieldView: View {
         .onDebouncedChange(of: $usernameEmailInput, debounceFor: 0.1) { newValue in
             showingTwoFactorField = false
             showingLoginButtonWarning = false
-            usernameEmailInvalid = !ValidationUtils.isValidUsername(input: newValue) && !ValidationUtils.isValidEmail(input: newValue)
+            usernameEmailInvalid = !ValidationUtils.isValidUsername(input: newValue) &&
+            !ValidationUtils.isValidEmail(input: newValue)
         }
     }
 }
