+++ Lunar/MockData/MockData.swift
@@ -269,4 +269,10 @@ struct MockData {
     static let userInfoView3 = Creator(id: 987_654, name: "Bookworm_27", displayName: Optional("Bookworm_27"), avatar: Optional("https://travelersworld.com/media/39/1e/391e4c07c13b4b87197065a8e9eb7f60485dd08c0a8e788a7ad6c82473e7a7be.jpg"), banned: false, published: "2023-07-25T12:30:55", actorID: "https://travelersworld.com/u/Bookworm_27", bio: Optional("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Aliquet bibendum enim facilisis gravida. Tincidunt ornare massa eget egestas purus viverra. Non sodales neque sodales ut etiam sit amet nisl purus. Sit amet consectetur adipiscing elit pellentesque habitant morbi. Sed risus pretium quam vulputate dignissim suspendisse in est. Lobortis scelerisque fermentum dui faucibus. Tincidunt vitae semper quis lectus nulla at volutpat. Velit euismod in pellentesque massa placerat duis ultricies. Penatibus et magnis dis parturient montes nascetur. Tincidunt vitae semper quis lectus. Aliquam sem fringilla ut morbi tincidunt augue interdum velit euismod. Urna neque viverra justo nec ultrices dui. At varius vel pharetra vel turpis nunc. Porttitor eget dolor morbi non arcu risus quis varius quam. Adipiscing elit duis tristique sollicitudin nibh sit amet commodo. Nisi quis eleifend quam adipiscing vitae proin sagittis. Sapien pellentesque habitant morbi tristique. Nisi vitae suscipit tellus mauris a diam maecenas sed enim. Congue nisi vitae suscipit tellus mauris. Dui sapien eget mi proin sed libero enim. Arcu cursus euismod quis viverra nibh cras. Arcu ac tortor dignissim convallis aenean et tortor at risus. Orci sagittis eu volutpat odio facilisis. Nulla facilisi etiam dignissim diam quis. Donec enim diam vulputate ut pharetra sit amet aliquam id. Mauris vitae ultricies leo integer malesuada nunc vel risus commodo. Pellentesque habitant morbi tristique senectus et netus et malesuada."), local: true, banner: Optional("https://travelersworld.com/media/80/23/8023bbab2d1a4d1c3b6e11401259b2df6497632e387c69cda7c74b1e41b9ce00.jpg"), deleted: false, admin: false, botAccount: false, instanceID: 205, updated: nil, matrixUserID: nil)
 
     static let userInfoView4 = Creator(id: 650_713, name: "maniel", displayName: nil, avatar: Optional("https://lemmy.ml/pictrs/image/3c00cbaa-f2d4-4b16-88bf-2438db3ab9c4.jpeg"), banned: false, published: "2023-06-07T09:24:31.398108", actorID: "https://lemmy.ml/u/maniel", bio: nil, local: true, banner: nil, deleted: false, admin: false, botAccount: false, instanceID: 394, updated: nil, matrixUserID: nil)
+
+    static let commentsViewPostTitle1 = "Lemmy.world active users is tapering off while other servers are gaining serious traction."
+    static let commentsViewThumbnailURL1 = "https://www.wallpapers13.com/wp-content/uploads/2015/12/Nature-Lake-Bled.-Desktop-background-image-1680x1050.jpg"
+    static let commentsViewPostBody1 = """
+            I noticed my feed on Lemmy was pretty dry today, even for Lemmy. Took me a while to realize lemmy.ml has been going up and down all morning
+            """
 }
