+++ Lunar/Tools/InsertSorter.swift
@@ -12,7 +12,8 @@ class InsertSorter {
         var index = comments.endIndex
         for (currentIndex, existingComment) in comments.enumerated() {
             if newComment.comment.path < existingComment.comment.path ||
-                (newComment.comment.path == existingComment.comment.path && newComment.comment.id < existingComment.comment.id) {
+                (newComment.comment.path == existingComment.comment.path &&
+                 newComment.comment.id < existingComment.comment.id) {
                 index = currentIndex
                 break
             }
