+++ Lunar/Tools/ViewExtension.swift
@@ -17,14 +17,12 @@ extension View {
             impact.impactOccurred()
         }
     }
-
     func hapticNotificationFeedbackOnTap(style: UINotificationFeedbackGenerator.FeedbackType) -> some View {
         onTapGesture {
             let haptic = UINotificationFeedbackGenerator()
             haptic.notificationOccurred(style)
         }
     }
-
     func onDebouncedChange<V>(
         of binding: Binding<V>,
         debounceFor dueTime: TimeInterval,
@@ -32,7 +30,6 @@ extension View {
     ) -> some View where V: Equatable {
         modifier(ListenDebounce(binding: binding, dueTime: dueTime, action: action))
     }
-
     /// Conditional Modifier
     /// **Usage:**
     /// ```
@@ -56,13 +53,11 @@ private struct ListenDebounce<Value: Equatable>: ViewModifier {
     @StateObject
     var debounceSubject: ObservableDebounceSubject<Value, Never>
     let action: (Value) -> Void
-
     init(binding: Binding<Value>, dueTime: TimeInterval, action: @escaping (Value) -> Void) {
         _binding = binding
         _debounceSubject = .init(wrappedValue: .init(dueTime: dueTime))
         self.action = action
     }
-
     func body(content: Content) -> some View {
         content
             .onChange(of: binding) { value in
@@ -74,27 +69,23 @@ private struct ListenDebounce<Value: Equatable>: ViewModifier {
     }
 }
 
-private final class ObservableDebounceSubject<Output: Equatable, Failure>: Subject, ObservableObject where Failure: Error {
+private final class ObservableDebounceSubject<Output: Equatable,
+                                              Failure>: Subject,
+                                                        ObservableObject where Failure: Error {
     private let passthroughSubject = PassthroughSubject<Output, Failure>()
-
     let dueTime: TimeInterval
-
     init(dueTime: TimeInterval) {
         self.dueTime = dueTime
     }
-
     func send(_ value: Output) {
         passthroughSubject.send(value)
     }
-
     func send(completion: Subscribers.Completion<Failure>) {
         passthroughSubject.send(completion: completion)
     }
-
     func send(subscription: Subscription) {
         passthroughSubject.send(subscription: subscription)
     }
-
     func receive<S>(subscriber: S) where S: Subscriber, Failure == S.Failure, Output == S.Input {
         passthroughSubject
             .removeDuplicates()
@@ -108,19 +99,15 @@ public extension View {
     // to convert the newly-made UIView to a UIImage.
     func asUIImage() -> UIImage {
         let controller = UIHostingController(rootView: self)
-
         // Set the background to be transparent incase the image is a PNG, WebP or (Static) GIF
         controller.view.backgroundColor = .black
-
         controller.view.frame = CGRect(x: 0, y: CGFloat(Int.max), width: 1, height: 1)
         UIApplication.shared.connectedScenes.flatMap { ($0 as? UIWindowScene)?.windows ?? [] }
             .filter(\.isKeyWindow)
             .first!.rootViewController?.view.addSubview(controller.view)
-
         let size = controller.sizeThatFits(in: UIScreen.main.bounds.size)
         controller.view.bounds = CGRect(origin: .zero, size: size)
         controller.view.sizeToFit()
-
         // here is the call to the function that converts UIView to UIImage: `.asUIImage()`
         let image = controller.view.asUIImage()
         controller.view.removeFromSuperview()
